+++ NEWS.md
@@ -1,5 +1,7 @@
 - Version 0.11.0 (unreleased)
   - `--clipboard` option removed in favour of `--clipboard-cmd`
+  - `--dmenu-prompt` option and trailing dmenu arguments removed in favour of
+    `--menu-cmd`
 
 - Version 0.10.0 - released 2020-08-02
   - Added support for arbitrary clipboard commands - thanks @ign0tus!